requirements.txt
flask
gunicorn
pandas
numpy
plotly
openpyxl
requests
cohere
python-dotenv